+++ src/settings.ts
@@ -27,7 +27,6 @@
 module powerbi.extensibility.visual {
     // powerbi.extensibility.utils.dataview
     import DataViewObjectsParser = utils.dataview.DataViewObjectsParser;
-    export type DateType = "Second" | "Minute" | "Hour" | "Day" | "Week" | "Month" | "Year";
 
     export class GanttSettings extends DataViewObjectsParser {
         general: GeneralSettings = new GeneralSettings();
@@ -80,7 +79,7 @@ module powerbi.extensibility.visual {
 
     export class DateTypeSettings {
         // tslint:disable-next-line:no-reserved-keywords
-        type: DateType = "Week";
+        type: DateTypes = DateTypes.Week;
         todayColor: string = "#000000";
         axisColor: string = "#000000";
         axisTextColor: string = "#000000";