+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 1.5.1
+ * Fixed tooltip label and date format of milestone line
 ## 1.5.0
  * Added ability to use custom date format for tooltip dates
  * Added ability to use extra information in the tooltip